requirements.txt
Flask
yfinance
pandas
numpy
ta
python-dotenv
gunicorn
